@@ -2,6 +2,7 @@
   "name": "@lapo/asn1js",
   "version": "1.3.0",
   "description": "Generic ASN.1 parser/decoder that can decode any valid ASN.1 DER or BER structures.",
+  "type": "module",
   "main": "asn1.js",
   "repository": {
     "type": "git",
@@ -20,7 +21,7 @@
     "test": "node test"
   },
   "engines": {
-    "node": ">=6.4.0"
+    "node": ">=12.20.0"
   },
   "devDependencies": {
     "eslint": "^8.34.0"
@@ -28,9 +29,12 @@
   "eslintConfig": {
     "env": {
       "es6": true,
-      "amd": true,
       "browser": true,
       "node": true
+    },
+    "parserOptions": {
+      "ecmaVersion": 2015,
+      "sourceType": "module"
     },
     "extends": [ "eslint:recommended" ],
     "globals": {
@@ -38,7 +42,7 @@
     },
     "rules": {
       "strict": [ "error", "function" ],
-      "indent": [ "error", 4, { "ignoredNodes": [ "Program > ExpressionStatement > CallExpression > FunctionExpression > BlockStatement > ExpressionStatement[directive='use strict']:first-child" ] } ],
+      "indent": [ "error", 4 ],
       "linebreak-style": [ "error", "unix" ],
       "semi": [ "warn", "always" ],
       "quotes": [ "error", "single", { "avoidEscape": true } ],
@@ -66,6 +70,11 @@
         "rules": {
           "comma-dangle": [ "error", "always-multiline" ],
           "quotes": [ "warn", "double" ]
+        }
+      }, {
+        "files": [ "defs.js" ],
+        "parserOptions": {
+          "ecmaVersion": 2021
         }
       }
     ]
